Move justification generation into test_helpers
