+++ primitives/substrate/src/lib.rs
@@ -17,6 +17,4 @@
 #![cfg_attr(not(feature = "std"), no_std)]
 
 pub mod justification;
-
-// #[cfg(test)]
 pub mod test_helpers;
